Create Get Possible Property
My scenario was a folder that included many grouped-objects. Each object included either a tool or a part that wasn't tooled. This snippet allows us to determine tool or part without causing an error when it is a tool (doesn't have a position property). 

Keywords: "Is Value Set", "Does Value Exist", "Determine if has the attribute", "Does object have property"